Perormance improvements
Code Cleanup
