+++ extensions/nextjs-docs/CHANGELOG.md
@@ -1,3 +1,7 @@
 # nextjs Changelog
 
+## [Performance Improvement] - 2022-04-30
+- Performance improvement
+- Caching improvement
+
 ## [Initial Version] - 2022-04-06
