+++ extensions/nextjs-docs/src/index.tsx
@@ -1,23 +1,37 @@
 import { ActionPanel, Action, Icon, List, showToast, Toast } from "@raycast/api";
-import TopicDetail from "./TopicDetail";
-import { Topic } from "./types/GithubType";
+import Topic from "./Topic";
+import { TopicType } from "./types/GithubType";
 import { useEffect, useState } from "react";
-import { getTopicsFromCache as getTopics } from "./services/Github";
+import { getPagesFromCache, checkForUpdates } from "./services/NextjsPages";
 
 export default function main() {
-  const [topics, setTopics] = useState<Topic[]>();
+  const [topics, setTopics] = useState<TopicType[]>();
 
   useEffect(() => {
-    getTopics()
-      .then((results: string) => {
-        setTopics(JSON.parse(results));
-      })
-      .catch((err: string) => {
-        showToast({
-          style: Toast.Style.Failure,
-          title: "Sorry, something went wrong! " + err,
-        });
+    async function getAllPages() {
+      const toast = await showToast({
+        style: Toast.Style.Animated,
+        title: "Fetching Data...",
       });
+
+      const cached_pages = await getPagesFromCache().catch((err) => {
+        toast.style = Toast.Style.Failure;
+        toast.title = "Let's try fetching fresh data...";
+        toast.show();
+      });
+
+      if (cached_pages) {
+        setTopics(JSON.parse(cached_pages));
+      }
+
+      const updated_data = await checkForUpdates();
+      if (updated_data) {
+        setTopics(JSON.parse(updated_data));
+      }
+
+      toast.hide();
+    }
+    getAllPages();
   }, []);
 
   if (!topics) {
@@ -33,7 +47,7 @@ export default function main() {
           title={topic.title}
           actions={
             <ActionPanel>
-              <Action.Push title={`Browse ${topic.title}`} target={<TopicDetail topic={topic} />} />
+              <Action.Push title={`Browse ${topic.title}`} target={<Topic topic={topic} />} />
             </ActionPanel>
           }
           accessories={[{ text: topic.filepath }]}
